mentor-hub/requirements.txt
fastapi
uvicorn
sqlalchemy
scikit-learn
pydantic
werkzeug
flask
pandas
flask-login
flask-sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
python-multipart